requirements.txt
python-telegram-bot
requests
python-dotenv
pandas
schedule
numpy
python-dateutil
cachetools